Moved all logic regarding the confirm dialog from the yasgui-component directive to the sparql view controller. The directive is used in other views such as "SPARQL Templates" and "Resources". In these other views, the plugins are not displayed, making this functionality unnecessary.
